Refactor/#27 question set page update (#45)
* Filter icon

There is a new filter icon after element of search input (homepage). The icon is not on dark background and we agreed on the default light background because the search box will change in the future for search customization improvement.  The search box is aligned rights.

* Fix title

The title was not showing in the small card component.

* Centred pagination

The pagination below list  is centred (homepage).

* Categories maping

A selection of distinct categories should be solved on the BE side. The response model will include a list of categories for mapping on specified tags.

* API update

Missing delete operation for QuestionSet

* Delete QuestionSetList

The context menu includes item "delete" with a defined callback function. The QuestionList is extended by delete operation on each QuestionSet.

* Remove Link element

Delete operation is not using any redirection, so it is not necessary to nest a link element into the button.

* Modal window

There is showing a confirmation window before delete and after a successful delete operation is appearing the second modal window with success info.

* TagList component

A list of categories (tags) are situated in small and large QuestionListCard. Rendering tags is resolved in a new component.